circleci: run make shorttest on windows. flag off a large swath of broken tests
