+++ internal/engine/build_and_deployer_live_update_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package engine
 
 import (
