+++ internal/engine/configs/configs_controller_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package configs
 
 import (
