+++ internal/engine/k8srollout/podmonitor_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package k8srollout
 
 import (
