+++ internal/engine/live_update_build_and_deployer_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package engine
 
 import (
