+++ internal/engine/local/controller_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package local
 
 import (
