+++ internal/engine/local/execer_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package local
 
 import (
