+++ internal/engine/pod_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package engine
 
 import (
