+++ internal/engine/telemetry/controller_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package telemetry
 
 import (
