+++ internal/hud/webview/helpers_test.go
@@ -1,0 +1,19 @@
+package webview
+
+import (
+	"github.com/windmilleng/wmclient/pkg/analytics"
+
+	"github.com/windmilleng/tilt/internal/store"
+	"github.com/windmilleng/tilt/pkg/model"
+)
+
+func newState(manifests []model.Manifest) *store.EngineState {
+	ret := store.NewState()
+	ret.AnalyticsEnvOpt = analytics.OptDefault
+	for _, m := range manifests {
+		ret.ManifestTargets[m.Name] = store.NewManifestTarget(m)
+		ret.ManifestDefinitionOrder = append(ret.ManifestDefinitionOrder, m.Name)
+	}
+
+	return ret
+}
