+++ internal/store/subscriber_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package store
 
 import (
@@ -145,62 +147,4 @@ func TestSubscriberTeardownOnRemove(t *testing.T) {
 		assert.Contains(t, err.Error(), "context canceled")
 	}
 	assert.Equal(t, 1, s.teardownCount)
-}
-
-type fakeSubscriber struct {
-	onChange      chan onChangeCall
-	setupCount    int
-	teardownCount int
-}
-
-func newFakeSubscriber() *fakeSubscriber {
-	return &fakeSubscriber{
-		onChange: make(chan onChangeCall),
-	}
-}
-
-type onChangeCall struct {
-	done chan bool
-}
-
-func (f *fakeSubscriber) assertOnChangeCount(t *testing.T, count int) {
-	t.Helper()
-
-	for i := 0; i < count; i++ {
-		f.assertOnChange(t)
-	}
-
-	select {
-	case <-time.After(50 * time.Millisecond):
-		return
-
-	case call := <-f.onChange:
-		close(call.done)
-		t.Fatalf("Expected only %d OnChange calls. Got: %d", count, count+1)
-	}
-}
-
-func (f *fakeSubscriber) assertOnChange(t *testing.T) {
-	t.Helper()
-
-	select {
-	case <-time.After(50 * time.Millisecond):
-		t.Fatalf("timed out waiting for subscriber.OnChange")
-	case call := <-f.onChange:
-		close(call.done)
-	}
-}
-
-func (f *fakeSubscriber) OnChange(ctx context.Context, st RStore) {
-	call := onChangeCall{done: make(chan bool)}
-	f.onChange <- call
-	<-call.done
-}
-
-func (f *fakeSubscriber) SetUp(ctx context.Context) {
-	f.setupCount++
-}
-
-func (f *fakeSubscriber) TearDown(ctx context.Context) {
-	f.teardownCount++
 }
