+++ internal/tiltfile/encoding/json_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package encoding
 
 import (
