+++ internal/tiltfile/git/git_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package git
 
 import (
