+++ internal/tiltfile/include/include_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package include
 
 import (
