+++ internal/tiltfile/include_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package tiltfile
 
 import (
