+++ internal/tiltfile/io/io_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package io
 
 import (
