+++ internal/tiltfile/os/os_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package os
 
 import (
