+++ internal/tiltfile/telemetry/telemetry_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package telemetry
 
 import (
