+++ internal/tiltfile/tiltextension/store_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package tiltextension
 
 import (
