+++ internal/watch/watcher_naive_test.go
@@ -1,4 +1,4 @@
-// +build !darwin
+// +build !darwin,!windows
 
 package watch
